Explicit conversions to make Clang happy

+++ src/core/field_types/core_uint_type.cpp
@@ -5,5 +5,5 @@ using namespace rive;
 
 unsigned int CoreUintType::deserialize(BinaryReader& reader)
 {
-	return reader.readVarUint();
+	return (int) reader.readVarUint();
 }
